Add support for more device info
USB devices frequently change their mount points on system reboot,
causing faulty job associations to drives. To prevent this, drive
recognition based on pyudev serial information (maker, model, serial
number) of optical drives is introduced.

This allows us to always associate a physical drive to the drives stored
in the database without relying on fixed /dev/sr* numbers. On system
startup, drives are now always scanned to recognize hardware changes.
For backward compatibility, the stored serial is used primarily to
identify the drive, and the mount point is only used as a fallback if
the identification fails. This change also makes the drive name
non-editable by the user. We are now also able to mark stale drives that
were previously attached but are not found by the current pyudev run.

This commit also addresses the unreliable eject status that was stored
in the database. - If the drive was ejected (e.g. before system start)
without the database being updated, this drive status became outdated.

We have also added an image to illustrate that a drive was not found and
was marked as stale in the settings overview. We reworked the Drive
Settings Overview to use collapsible accordion items of bootstrap card
items. The icons are now (to some extend) scalable instead of fixed
size. The layout of the drive tab is now 2 column based to give the icon
some space to extend to the bottom.

This commit also addresses some of the pylint errors, introduces
dataclasses for pyudev object handling and aims for better code
readability.