Add ripping finished ready criterion for drives
The drive status is only updated for "success" and "fail" status.
However, the drive is free to use after the Ripping process finished.
This is currently whenever we go to the transcoding step. Some rework of
the state logic is needed in v3 but this is good for now.

Issue #1236